requirements.txt
sentence-transformers
faiss-cpu
langdetect
streamlit
requests
